# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface InspectDownloadInput {
  /** @minLength 8 */
  url: string;
}

export type CreateDownloadInputMediaType = typeof CreateDownloadInputMediaType[keyof typeof CreateDownloadInputMediaType];


export const CreateDownloadInputMediaType = {
  video: 'video',
  audio: 'audio',
} as const;

export interface CreateDownloadInput {
  /** @minLength 8 */
  url: string;
  /** @minLength 1 */
  formatId: string;
  mediaType?: CreateDownloadInputMediaType;
}

export type DownloadPlatform = typeof DownloadPlatform[keyof typeof DownloadPlatform];


export const DownloadPlatform = {
  youtube: 'youtube',
  instagram: 'instagram',
  threads: 'threads',
  tiktok: 'tiktok',
  unknown: 'unknown',
} as const;

export type DownloadFormatKind = typeof DownloadFormatKind[keyof typeof DownloadFormatKind];


export const DownloadFormatKind = {
  video: 'video',
  audio: 'audio',
} as const;

export interface DownloadFormat {
  id: string;
  label: string;
  extension: string;
  kind: DownloadFormatKind;
  /** @nullable */
  sizeLabel?: string | null;
}

export interface DownloadInspection {
  url: string;
  platform: DownloadPlatform;
  isSupported: boolean;
  /** @nullable */
  title: string | null;
  /** @nullable */
  thumbnailUrl: string | null;
  /** @nullable */
  durationSeconds: number | null;
  formats: DownloadFormat[];
}

export type DownloadJobStatus = typeof DownloadJobStatus[keyof typeof DownloadJobStatus];


export const DownloadJobStatus = {
  queued: 'queued',
  downloading: 'downloading',
  completed: 'completed',
  failed: 'failed',
} as const;

export type DownloadJobMediaType = typeof DownloadJobMediaType[keyof typeof DownloadJobMediaType];


export const DownloadJobMediaType = {
  video: 'video',
  audio: 'audio',
} as const;

export interface DownloadJob {
  id: string;
  url: string;
  platform: DownloadPlatform;
  /** @nullable */
  title: string | null;
  status: DownloadJobStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  mediaType: DownloadJobMediaType;
  /** @nullable */
  filename: string | null;
  createdAt: string;
  /** @nullable */
  downloadUrl: string | null;
  /** @nullable */
  error: string | null;
}