# lib/api-zod/src/generated/types/createDownloadInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateDownloadInputMediaType } from './createDownloadInputMediaType';

export interface CreateDownloadInput {
  /** @minLength 8 */
  url: string;
  /** @minLength 1 */
  formatId: string;
  mediaType?: CreateDownloadInputMediaType;
}