# lib/api-zod/src/generated/types/downloadFormat.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DownloadFormatKind } from './downloadFormatKind';

export interface DownloadFormat {
  id: string;
  label: string;
  extension: string;
  kind: DownloadFormatKind;
  /** @nullable */
  sizeLabel?: string | null;
}