# lib/api-zod/src/generated/types/downloadInspection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DownloadFormat } from './downloadFormat';
import type { DownloadPlatform } from './downloadPlatform';

export interface DownloadInspection {
  url: string;
  platform: DownloadPlatform;
  isSupported: boolean;
  /** @nullable */
  title: string | null;
  /** @nullable */
  thumbnailUrl: string | null;
  /** @nullable */
  durationSeconds: number | null;
  formats: DownloadFormat[];
}