# lib/api-zod/src/generated/types/downloadJob.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DownloadJobMediaType } from './downloadJobMediaType';
import type { DownloadJobStatus } from './downloadJobStatus';
import type { DownloadPlatform } from './downloadPlatform';

export interface DownloadJob {
  id: string;
  url: string;
  platform: DownloadPlatform;
  /** @nullable */
  title: string | null;
  status: DownloadJobStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  mediaType: DownloadJobMediaType;
  /** @nullable */
  filename: string | null;
  createdAt: string;
  /** @nullable */
  downloadUrl: string | null;
  /** @nullable */
  error: string | null;
}