# lib/api-zod/src/generated/types/downloadPlatform.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type DownloadPlatform = typeof DownloadPlatform[keyof typeof DownloadPlatform];


export const DownloadPlatform = {
  youtube: 'youtube',
  instagram: 'instagram',
  threads: 'threads',
  tiktok: 'tiktok',
  unknown: 'unknown',
} as const;